refactor: split results.ts into focused sub-modules
Extract results-analysis.ts (trajectory, correlation, insights, context)
and results-query.ts (diff, status, trend, baseline) from the 772-line
results.ts. Core CRUD stays in results.ts (116 lines). All 414 tests
pass unchanged — no behavior changes.

Closes #110
