+++ .canductor/results.tsv
@@ -40,3 +40,4 @@ guard-ref	2026-03-23T23:13:27.937Z	38	block	typecheck:100,tests:0,code_quality:0
 104	2026-03-24T06:13:12.819Z	98	auto_merge	typecheck:100,tests:100,code_quality:93	merged	Verified 104
 108	2026-03-24T07:07:35.285Z	99	auto_merge	typecheck:100,tests:100,code_quality:94	merged	Verified 108
 109	2026-03-24T07:10:34.722Z	98	auto_merge	typecheck:100,tests:100,code_quality:93	merged	Verified 109
+110	2026-03-24T07:15:38.185Z	98	auto_merge	typecheck:100,tests:100,code_quality:92	pending	Verified 110
