+++ packages/core/src/index.ts
@@ -1,8 +1,10 @@
 export { loadConfig, findConfigPath, writeConfigBaseline, validateConfig, listLayers } from './config.js';
 export { verify, computeCompositeScore, evaluatePolicy } from './verify.js';
 export { runLayer, runDeterministicLayer, runScreenshotDiffLayer, runAgentReviewLayer, getAgentReviewPrompt } from './layers.js';
-export { readResults, appendResult, updateResultStatus, analyzeResults, detectStalls, correlateLayerFailures, analyzeTrajectory, generateInsights, generatePromptContext, diffResults, getStatus, getTrend, computeAutoBaseline, getBaseline } from './results.js';
-export type { LayerDiff, DiffResult, PipelineStatus, TrendEntry, TrendResult } from './results.js';
+export { readResults, appendResult, updateResultStatus } from './results.js';
+export { analyzeResults, detectStalls, correlateLayerFailures, analyzeTrajectory, generateInsights, generatePromptContext } from './results-analysis.js';
+export { diffResults, getStatus, getTrend, computeAutoBaseline, getBaseline } from './results-query.js';
+export type { LayerDiff, DiffResult, PipelineStatus, TrendEntry, TrendResult } from './results-query.js';
 export { compareScreenshots, updateBaseline } from './screenshot.js';
 export type { ScreenshotDiffResult } from './screenshot.js';
 export { evaluateExpression, evaluateAllPolicies, buildPolicyContext } from './policy.js';
